+++ exemplos/diletta_app/android/app/src/main/kotlin/com/dilettasolutions/diletta_app/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.dilettasolutions.diletta_app
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()